Update city objects tree item
This changes the expand/collapse icon, fixes the width of the object
type icon and replace the way geometry badges are shown.

A new GeometryBadge component is introduced.

Caveats: there are performance concerns regarding the fact that the
previous v-for loop between geometries and geometry instances is not
converted to a v-if for every geometry inside the GeometryBadge.